requirements.txt
# AI Coach - Core Dependencies
# Python 3.11+

# RAG and Vector Search
faiss-cpu==1.8.0
numpy>=1.24.0,<2.0.0

# OpenAI-compatible client for NVIDIA NIMs
openai>=1.12.0

# LangGraph for Agent Orchestration
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0

# HTTP Client for Simulator API
httpx>=0.26.0
websockets>=12.0

# FastAPI for REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Processing
pydantic>=2.5.0
python-dotenv>=1.0.0

# Kubernetes Client (optional, for deployment scripts)
kubernetes>=28.1.0

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.9
